Extracts classes.
