+++ financial-connections/src/main/java/com/stripe/android/financialconnections/ui/components/Button.kt
@@ -249,88 +249,3 @@ internal fun FinancialConnectionsButtonPreview() {
         }
     }
 }
-
-internal interface MultipleEventsCutter {
-    fun processEvent(event: () -> Unit)
-
-    companion object
-}
-
-internal fun MultipleEventsCutter.Companion.get(): MultipleEventsCutter =
-    MultipleEventsCutterImpl()
-
-private class MultipleEventsCutterImpl : MultipleEventsCutter {
-    private val now: Long
-        get() = System.currentTimeMillis()
-
-    private var lastEventTimeMs: Long = 0
-
-    override fun processEvent(event: () -> Unit) {
-        if (now - lastEventTimeMs >= DEBOUNCE_MS) {
-            event.invoke()
-        }
-        lastEventTimeMs = now
-    }
-
-    companion object {
-        private const val DEBOUNCE_MS = 500L
-    }
-}
-
-fun Modifier.clickableSingle(
-    enabled: Boolean = true,
-    onClickLabel: String? = null,
-    role: Role? = null,
-    onClick: () -> Unit
-
-) = composed(
-    inspectorInfo = debugInspectorInfo {
-        name = "clickable"
-        properties["enabled"] = enabled
-        properties["onClickLabel"] = onClickLabel
-        properties["role"] = role
-        properties["onClick"] = onClick
-    },
-
-    factory = {
-        val multipleEventsCutter = remember { MultipleEventsCutter.get() }
-        Modifier.clickable(
-            interactionSource = remember { MutableInteractionSource() },
-            indication = LocalIndication.current,
-            enabled = enabled,
-            onClickLabel = onClickLabel,
-            onClick = { multipleEventsCutter.processEvent { onClick() } },
-            role = role,
-        )
-    }
-)
-
-fun Modifier.clickableSingle(
-    interactionSource: MutableInteractionSource,
-    indication: Indication?,
-    enabled: Boolean = true,
-    onClickLabel: String? = null,
-    role: Role? = null,
-    onClick: () -> Unit
-
-) = composed(
-    inspectorInfo = debugInspectorInfo {
-        name = "clickable"
-        properties["enabled"] = enabled
-        properties["onClickLabel"] = onClickLabel
-        properties["role"] = role
-        properties["onClick"] = onClick
-    },
-
-    factory = {
-        val multipleEventsCutter = remember { MultipleEventsCutter.get() }
-        Modifier.clickable(
-            interactionSource = interactionSource,
-            indication = indication,
-            enabled = enabled,
-            onClickLabel = onClickLabel,
-            onClick = { multipleEventsCutter.processEvent { onClick() } },
-            role = role,
-        )
-    }
-)
