+++ financial-connections/src/main/java/com/stripe/android/financialconnections/ui/components/MultipleEventsCutter.kt
@@ -1,0 +1,108 @@
+package com.stripe.android.financialconnections.ui.components
+
+import androidx.compose.foundation.Indication
+import androidx.compose.foundation.LocalIndication
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.composed
+import androidx.compose.ui.platform.debugInspectorInfo
+import androidx.compose.ui.semantics.Role
+
+/**
+ * A wrapper around [clickable] that prevents multiple clicks from being registered within a
+ * short time frame.
+ */
+internal interface MultipleEventsCutter {
+    fun processEvent(event: () -> Unit)
+
+    companion object
+}
+
+internal fun MultipleEventsCutter.Companion.get(): MultipleEventsCutter =
+    MultipleEventsCutterImpl()
+
+private class MultipleEventsCutterImpl : MultipleEventsCutter {
+    private val now: Long
+        get() = System.currentTimeMillis()
+
+    private var lastEventTimeMs: Long = 0
+
+    override fun processEvent(event: () -> Unit) {
+        if (now - lastEventTimeMs >= DEBOUNCE_MS) {
+            event.invoke()
+        }
+        lastEventTimeMs = now
+    }
+
+    companion object {
+        private const val DEBOUNCE_MS = 500L
+    }
+}
+
+/**
+ * A wrapper around [clickable] that prevents multiple clicks from being registered within a
+ * short time frame.
+ */
+internal fun Modifier.clickableSingle(
+    enabled: Boolean = true,
+    onClickLabel: String? = null,
+    role: Role? = null,
+    onClick: () -> Unit
+
+) = composed(
+    inspectorInfo = debugInspectorInfo {
+        name = "clickable"
+        properties["enabled"] = enabled
+        properties["onClickLabel"] = onClickLabel
+        properties["role"] = role
+        properties["onClick"] = onClick
+    },
+
+    factory = {
+        val multipleEventsCutter = remember { MultipleEventsCutter.get() }
+        Modifier.clickable(
+            interactionSource = remember { MutableInteractionSource() },
+            indication = LocalIndication.current,
+            enabled = enabled,
+            onClickLabel = onClickLabel,
+            onClick = { multipleEventsCutter.processEvent { onClick() } },
+            role = role,
+        )
+    }
+)
+
+/**
+ * A wrapper around [clickable] that prevents multiple clicks from being registered within a
+ * short time frame.
+ */
+internal fun Modifier.clickableSingle(
+    interactionSource: MutableInteractionSource,
+    indication: Indication?,
+    enabled: Boolean = true,
+    onClickLabel: String? = null,
+    role: Role? = null,
+    onClick: () -> Unit
+
+) = composed(
+    inspectorInfo = debugInspectorInfo {
+        name = "clickable"
+        properties["enabled"] = enabled
+        properties["onClickLabel"] = onClickLabel
+        properties["role"] = role
+        properties["onClick"] = onClick
+    },
+
+    factory = {
+        val multipleEventsCutter = remember { MultipleEventsCutter.get() }
+        Modifier.clickable(
+            interactionSource = interactionSource,
+            indication = indication,
+            enabled = enabled,
+            onClickLabel = onClickLabel,
+            onClick = { multipleEventsCutter.processEvent { onClick() } },
+            role = role,
+        )
+    }
+)
